refactor: modify put article
- put 메소드 에러로 post로 글 수정 로직 변경
- ArticleDetailView Class로 view위치 변경
- 유저 권한 확인하는 로직 validator 에 추가
- 수정 기능 unit test 작성

+++ diary/services.py
@@ -2,6 +2,8 @@ import boto3
 import time
 from datetime import datetime
 
+from django.db.utils import IntegrityError
+
 from config.settings import AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, AWS_STORAGE_BUCKET_NAME
 
 from diary.dto    import ArticleCreateDto, ArticleIdDto, ArticleUpdateDto
@@ -19,7 +21,7 @@ class ArticleService():
 
     @staticmethod
     def filter_article(dto:ArticleIdDto):
-        return Article.objects.filter(id = dto.id).values("id", "title", "content", "image", "created_at", "updated_at")
+        return Article.objects.filter(id=dto.id).values("id", "title", "content", "image", "created_at", "updated_at")
 
     @staticmethod
     def create(dto:ArticleCreateDto, url):
@@ -45,49 +47,61 @@ class ArticleService():
         )
         Article.objects.filter(id = dto.id).delete()
 
+    @staticmethod
     def update(dto:ArticleIdDto, file):
-        if file:
-            Article.objects.filter(id = dto.id).update(
-            title = dto.title,
-            content = dto.content,
-            updated_at = time.time(),
-            image =  file
-            )
-        else:
-            Article.objects.filter(id = dto.id).update(
-            title = dto.title,
-            content = dto.content,
-            updated_at = time.time(),
-            )      
+        try:
+            if file:
+                Article.objects.filter(id=dto.id).update(
+                title=dto.title,
+                content=dto.content,
+                updated_at=time.time(),
+                image=file
+                )
+
+            else:
+                Article.objects.filter(id=dto.id).update(
+                title=dto.title,
+                content=dto.content,
+                updated_at=time.time(),
+                )      
+
+            return {"error": False}
+
+        except IntegrityError:            
+            return {"error": True}
 
 
 class UploadImageService():
     def upload(dto:ArticleUpdateDto):
         s3_url = "https://django-diary-bucket.s3.ap-northeast-2.amazonaws.com/"
         
-        client = boto3.client(
-            "s3",
-            aws_access_key_id=AWS_ACCESS_KEY_ID,
-            aws_secret_access_key=AWS_SECRET_ACCESS_KEY,
-        )
-        
-        bucket = AWS_STORAGE_BUCKET_NAME
-        
-        if "id" in dto.__dict__: # 이미지 업데이트 경우
-            client.delete_object(
-                Bucket=bucket,
-                Key=Article.objects.get(id=dto.id).image.split("/")[-1]
+        if not dto.file:
+            image_name = Article.objects.get(id=dto.id).image
+            
+        else:
+            client = boto3.client(
+                "s3",
+                aws_access_key_id=AWS_ACCESS_KEY_ID,
+                aws_secret_access_key=AWS_SECRET_ACCESS_KEY,
             )
+            
+            bucket = AWS_STORAGE_BUCKET_NAME
+            
+            if "id" in dto.__dict__: # 이미지 업데이트 경우
+                client.delete_object(
+                    Bucket=bucket,
+                    Key=Article.objects.get(id=dto.id).image.split("/")[-1]
+                )
 
-        now = datetime.now().strftime("%Y%H%M%S")
-        
-        client.upload_fileobj(
-            dto.file,
-            bucket,
-            now+dto.file.name,
-            ExtraArgs={"ContentType": dto.file.content_type,}
-            )
-        
-        return s3_url+now+dto.file.name
+            now = datetime.now().strftime("%Y%H%M%S")
+            
+            client.upload_fileobj(
+                dto.file,
+                bucket,
+                now+dto.file.name,
+                ExtraArgs={"ContentType": dto.file.content_type,}
+                )
+            image_name = s3_url+now+dto.file.name
+        return image_name
